pr tweaks
