+++ anotamela/__init__.py
@@ -1,6 +1,7 @@
 from os import environ
 import logging
 
+from Bio import Entrez
 import coloredlogs
 
 from .annotators import *
@@ -12,4 +13,6 @@ coloredlogs.install(level=environ.get('LOGLEVEL') or 'INFO')
 
 logging.getLogger('requests').setLevel(logging.WARNING)
 logging.getLogger('urllib3').setLevel(logging.WARNING)
+
+Entrez.tool = 'anotamela'
