+++ anotamela/helpers/helpers.py
@@ -1,7 +1,9 @@
+from os.path import expanduser
 from itertools import zip_longest
 from functools import lru_cache
 
 from bs4 import BeautifulSoup
+from Bio import Entrez
 
 
 def grouped(iterable, group_size):
@@ -24,4 +26,16 @@ def listify(maybe_list):
     if isinstance(maybe_list, list):
         return maybe_list
     return [maybe_list]
+
+
+def set_email_for_entrez():
+    email_filepath = expanduser('~/.mail_address_for_Entrez')
+
+    try:
+        with open(email_filepath) as f:
+            Entrez.email = f.read().strip()
+    except FileNotFoundError:
+        msg = ('Please set a mail for Entrez in {}. Entrez will notify '
+                'that mail before banning you if your usage is too high.')
+        raise FileNotFoundError(msg.format(email_filepath))
